move shared embed interfaces into the "constants" packages
