+++ packages/constants/src/core/index.ts
@@ -1,0 +1,3 @@
+export type ElementType<T extends unknown[]> = T[number];
+
+export type ArrayMap = Record<string, unknown[]>;
